New README and changes to post_rake()
Started a new README using RMarkdown, the old one is still in use at the moment. Added the weighted n's to the output in post_rake().

We need to have consistent naming of things (my fault). I vote we use "." in argument names and data output (thinking pre/post_rake()), "_" for function names